+++ example/src/pages/index.js
@@ -5,7 +5,7 @@ const HomePage = () => {
     <main>
       <h1>Example Gatsby Site</h1>
       <p>This site demonstrates the gatsby-plugin-sitemap-html plugin.</p>
-      <a href="/sitemap.xml">View Sitemap</a>
+      <a href="/sitemap-index.xml">View Sitemap</a>
     </main>
   );
 };