+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Removed
+
+- Renaming `sitemap-index.xml` to `sitemap.xml`. Sitemap filenames now match [`gatsby-plugin-sitemap`](https://www.gatsbyjs.com/plugins/gatsby-plugin-sitemap/) output (typically `sitemap-index.xml`, `sitemap-0.xml`, …).
+
 ## [1.2.2] - 2025-11-07
 
 ### Changed